<fix>[host]: enhance PCI device vendor detection with vendor_id
Add vendor_id parameter to _simplify_pci_device_name method.
Add support for Haiguang vendor detection using.
both name and vendor_id.
Improve vendor identification reliability for
specific hardware.

Resolves: ZSTAC-78450

(cherry picked from commit f02a5378d3ef648ae794bf95a33b2c98fcecaef1)
