+++ zstacklib/zstacklib/utils/gpu.py
@@ -4,6 +4,8 @@ from zstacklib.utils import log, linux
 from zstacklib.utils.bash import *
 import json
 
+from zstacklib.utils.pci import VendorEnum
+
 logger = log.get_logger(__name__)
 
 
@@ -303,9 +305,5 @@ def is_valid_video_controller(device):
     return all(keyword not in device for keyword in invalid_keywords)
 
 
-def is_valid_co_processor(description, vendor_id):
-    description_keywords = {"Haiguang"}
-    valid_vendor_ids = {"1d94"}
-
-    return (any(keyword.lower() in description.lower() for keyword in description_keywords) or
-            vendor_id.lower() in valid_vendor_ids)
+def is_valid_co_processor(vendor):
+    return vendor in [VendorEnum.HAIGUANG]
