+++ zstacklib/zstacklib/utils/pci.py
@@ -4,11 +4,27 @@ import platform
 from zstacklib.utils import log, linux, sizeunit
 import xml.etree.ElementTree as ET
 
+from zstacklib.utils.bash import bash_roe
+
 logger = log.get_logger(__name__)
+
+
+class VendorEnum:
+    INTEL = "Intel"
+    AMD = "AMD"
+    NVIDIA = "NVIDIA"
+    HAIGUANG = "Haiguang"
+    HUAWEI = "Huawei"
+    TIANSHU = "TianShu"
+    VASTAI = "Vastai"
 
 
 def is_gpu(type):
     return type in ['GPU_3D_Controller', 'GPU_Video_Controller', 'GPU_Processing_Accelerators', 'GPU_Co_Processor']
+
+
+def is_haiguang_pci_device(vendor_id):
+    return vendor_id in ['1d94']
 
 
 def fmt_pci_address(pci_device):
@@ -168,6 +184,7 @@ def parse_resources(device_path):
     logger.info("get pci device[path: %s],resources: %s" % (device_path, resources))
     return resources
 
+
 def get_pci_passthrough_mapping(vm_dom):
     pci_mapping = {}
     xml_tree = ET.fromstring(vm_dom.XMLDesc())
@@ -188,3 +205,13 @@ def get_pci_passthrough_mapping(vm_dom):
         pci_mapping[vm_pci_address] = host_pci_address
 
     return pci_mapping
+
+
+def get_pci_device_ids():
+    # Get IDs using -Dmmnv (without second 'n' to avoid truncation)
+    return bash_roe("lspci -Dmmnv")
+
+
+def get_pci_device_names():
+    # Get names using -Dmmv (without 'nn' to get full names)
+    return bash_roe("lspci -Dmmv")
